+++ web/marker/point.js
@@ -16,10 +16,12 @@ class Point {
     }
 
     setupMarker() {
-        this.marker = L.marker([this.latitude, this.longitude], { draggable: Role.hasAtLeastRole('editor') })
+        let isEditor = Role.hasAtLeastRole('editor')
+
+        this.marker = L.marker([this.latitude, this.longitude], { draggable: isEditor })
 
         let lat, lng = (this.latitude, this.longitude)
-        if (Role.hasAtLeastRole('editor')) {
+        if (isEditor) {
             this.marker.on('drag', function (event) {
                 // for some reason this shit works, though should be event.originalEvent.button
                 if (event.originalEvent.buttons == 1) return
@@ -27,7 +29,7 @@ class Point {
                 event.target.setLatLng([lat, lng])
                 setTimeout(() => event.target.dragging.enable());
             })
-            
+
             let point = this
             let marker = this.marker
             this.marker.on('dragend', async function (event) {
@@ -43,19 +45,33 @@ class Point {
             })
         }
 
-        this.marker.bindPopup(
-            L.popup().setContent(
-                `
-                <div class="popup">
-                    Name: <input type="text" class="popupNameInput" id="${this.id}" maxlength="255" size="10" value="${this.name}"/><br/>
-                    Latitude: ${this.latitude.toFixed(4)}<br/>
-                    Longitude: ${this.longitude.toFixed(4)}
-                </div>
-                <button class="popupDeleteButton" onclick="Marker.shown.get(MarkerableTypes.Point).get(${this.id}).delete()">Delete</button>
-                <button class="popupUpdateButton" onclick="Marker.shown.get(MarkerableTypes.Point).get(${this.id}).pullUpdate()">Update</button>
-                `
-            )
-        ).openPopup()
+        if (isEditor) {
+            this.marker.bindPopup(
+                L.popup().setContent(
+                    `
+                    <div class="popup">
+                        Name: <input type="text" class="popupNameInput" id="${this.id}" maxlength="255" size="10" value="${this.name}"/><br/>
+                        Latitude: ${this.latitude.toFixed(4)}<br/>
+                        Longitude: ${this.longitude.toFixed(4)}
+                    </div>
+                    <button class="popupDeleteButton" onclick="Marker.shown.get(MarkerableTypes.Point).get(${this.id}).delete()">Delete</button>
+                    <button class="popupUpdateButton" onclick="Marker.shown.get(MarkerableTypes.Point).get(${this.id}).pullUpdate()">Update</button>
+                    `
+                )
+            ).openPopup()
+        } else {
+            this.marker.bindPopup(
+                L.popup().setContent(
+                    `
+                    <div class="popup">
+                        Name: <label class='popupNameInput' id='${this.id}' maxlength='255' size='10'>${this.name}</label><br/>
+                        Latitude: ${this.latitude.toFixed(4)}<br/>
+                        Longitude: ${this.longitude.toFixed(4)}
+                    </div>
+                    `
+                )
+            ).openPopup()
+        }
     }
 
     pullUpdate() {
@@ -72,7 +88,11 @@ class Point {
     async update(updateInfo) {
         updateInfo.id = this.id        
 
-        await MarkerFetch.updateType(MarkerableTypes.Point, updateInfo)
+        try {
+            MarkerFetch.updateType(MarkerableTypes.Point, updateInfo)
+        } catch (e) {
+            return
+        }
 
         if (updateInfo.name !== undefined) {
             this.name = updateInfo.name
